+++ client/components/app.jsx
@@ -1,12 +1,11 @@
-import React from 'react';
-import { Component } from 'react';
+import { React, Component } from 'react';
 
 export default class App extends Component {
   render() {
     return (
       <main>
         <h1>
-            Open Museum!            
+            Open Museum!
         </h1>
       </main>
     );
