--- client/routes.js
@@ -1,9 +1,0 @@
-import React from 'react'
-import {Route, IndexRoute, Router} from 'react-router'
-
-import App from './components/app.js'
-
-export default (
-  <Route path='/' component={App}/>
-)
-
