+++ client/routes.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { Route, IndexRoute, Router } from 'react-router';
+
+import App from './components/app';
+
+export default (
+  <Route path="/" component={App} />
+);
+
